完善

+++ XDr/Code/XDrive/Inc/control/adaptive_control.h
@@ -7,7 +7,7 @@ typedef struct
     u8 temp_u;
     u8 temp_v;
     u8 temp_w;
-    u8 tempareture;
+    float tempareture;
     float Udc;
 } ADAPTIVE_CON_T;
 extern ADAPTIVE_CON_T g_adaptive_con;
